Comparison is added
Here ins this time we have added two new classes that we use Update to update a word file usiing based on its excel file. and the Checker class is used to check for the differences betwwen rows of the excel and word document. if any differences is detected it will copy again the value from the excel document.
And Main class is used to get these classes to work.